Fixed ImportError when `Products.ATContentTypes` is not available.
This happens when you only have the `Products.CMFPlone` egg and not the `Plone` egg.
See https://community.plone.org/t/plone-5-1b3-soft-released/3936/5
Traceback was:

  File "/home/vagrant/Eggs/plone.app.upgrade-2.0.2-py2.7.egg/plone/app/upgrade/__init__.py", line 111, in <module>
      from Products.ATContentTypes.tool import metadata
   zope.configuration.xmlconfig.ZopeXMLConfigurationError: File "/home/vagrant/www.fhnw.ch/parts/instance1v/etc/site.zcml", line 12.2-12.39
     ZopeXMLConfigurationError: File "/home/vagrant/Eggs/Products.CMFPlone-5.1b3-py2.7.egg/Products/CMFPlone/meta.zcml", line 47.4-51.10
     ImportError: No module named ATContentTypes.tool

I fixed that with the already available atcontenttypes_bbb module.

Then the next traceback surfaced:

    ZopeXMLConfigurationError: File "/Users/maurits/community/plone-coredev/5.1/src/plone.app.upgrade/plone/app/upgrade/v40/configure.zcml", line 172.8-176.14
    ConfigurationError: ('Invalid value for', 'handler', "ImportError: Couldn't import Products.CMFDefault.upgrade.to22, No module named upgrade.to22")

This is because my fix turned Products.CMFDefault into an aliased module,
which caused the conditionally registered upgrade step to be loaded, causing an error.
I fixed this by making a wrapper around the original CMFDefault upgrade step.

Note that this will only be called when someone upgrades from a Plone 4.0 alpha release to Plone 5, which seems unlikely.