Hide our 'products' from installation for both CMFQuickInstallerTool and CMFPlone.
The plone.app.upgrade.v40 package, etc, would be shown in as installable
if you had a branch of CMFPlone with the QI code completely removed (scheduled for Plone 6)
but with the Products.CMFQuickInstallerTool package still available.

Now we register the HiddenProducts factory twice in that case, for both INonInstallable interfaces.