image scales catlog upgrade: only change the image_scales metadata.
Do not recalculate the other columns.
The new update_catalog_metadata function now accepts a column keyword argument.
If you use this, then we try to update only this column, using the inner workings of the catalog.
If we see the catalog is not what we expect, then we fall back to calling catalog_object, which touches all metadata.